Allow adding more table indexing configuration for Pinot tables.
Currently, we only allow the inverted indices to be added to Pinot table but
we need to be able to add other indexes like bloom filters, range indexes
to optimize the performance of Pinot queries. This PR let's that config to
be tunable easily.

I'm also rethinking if we should have this indirection in Pinot table config
or should we directly provide the JSON that's needed but that's for future PRs.

Also upgraded the pinot-tools to 0.5.0 version in this PR.